% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkType.R
\name{checkType}
\alias{checkType}
\title{checkType Function}
\usage{
checkType(count_matrix, n = 10)
}
\arguments{
\item{count_matrix}{The matrix that we want to check}

\item{n}{How many columns to check. Default to 10.}
}
\value{
the quantiles of the count matrix
}
\description{
Plot the boxplot from ten random columns of the input matrix. Useful to visualize
whether the provided count matrix is is logged or if any normalization was done.
}
\examples{
data <- matrix(rnbinom(1000 * 100, size = 1000, prob = .999), ncol = 100)
checkType(data)
}
