% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{clean Function}
\usage{
clean(tb)
}
\arguments{
\item{tb}{the \code{data.frame} or \code{tibble}}
}
\description{
This function remove columns of a dataframe that provide no info (identitcal accross all rows)
}
