% Generated by roxygen2: do not edit by hand
\name{colors}
\alias{colors}
\title{Shortcut for the RcolorBrewer function}
\usage{
colors(print = F)
}
\arguments{
\item{print}{If T, print the names of palettes in R Color Brewer}
}
\description{
A shortcut for the display.brewer.all() function of RColorBrewer
}
