% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{counts}
\title{counts Function}
\usage{
counts(count_matrix, n = 10)
}
\value{
the quantiles of the count matrix
}
\description{
Plot the boxplot from ten random columns of the input matrix. Useful to visualize
whether the provided count matrix is is logged or if any normalization was done.
}
