% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommonOrthologs.R
\name{.common}
\alias{.common}
\title{common Function}
\usage{
.common(gHuman, gMouse, gencodeHuman, gencodeMouse)
}
\arguments{
\item{gHuman}{The list of human genes}

\item{gMouse}{The list of mouse genes}
}
\description{
For the list of genes restrict to the common orthologs
}
