% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommonOrthologs.R
\name{.getGencode}
\alias{.getGencode}
\title{.getGencode Function}
\usage{
.getGencode(ref, species)
}
\arguments{
\item{ref}{the gencode reference for species.
Either a number (the gencode version) or the location of the file.}

\item{species}{The species}
}
\description{
Download and pre-process the gencode files
}
