% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommonOrthologs.R
\name{.Orthologs}
\alias{.Orthologs}
\title{.Orthologs Function}
\usage{
.Orthologs(genes, gencode, species)
}
\arguments{
\item{genes}{The list of genes for which we want the orthologs}

\item{gencode}{the gencode file}

\item{species}{The species}
}
\description{
For a list of genes, find the orthologs in the other species
}
