% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump_and_quit.R
\name{dump_and_quit}
\alias{dump_and_quit}
\title{dump_and_quit Function}
\usage{
dump_and_quit(file = "last.dump")
}
\value{
Invisible NULL
}
\description{
From https://adv-r.hadley.nz/debugging.html#alternatives
}
