% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipak.R
\name{ipak}
\alias{ipak}
\title{ipak Function}
\usage{
ipak(pkg)
}
\arguments{
\item{pcks}{a vector of characters naming packages}
}
\description{
This function loads the packages and install them from CRAN or Bioconductor if needed
}
