% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleJobMemoryUsage.R
\name{memoryUsage}
\alias{memoryUsage}
\title{memoryUsage Function}
\usage{
memoryUsage(loc)
}
\arguments{
\item{loc}{the file location}
}
\description{
Take as input a memory file usage and plot usage over time as well as maximum
 usage
}
