% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdf.R
\name{tdf}
\alias{tdf}
\title{Pipe compatible transpose}
\usage{
tdf(.data, rowname = NA, colname = "colnames")
}
\arguments{
\item{.data}{the data frame to transpose}

\item{rowname}{the name of the column that will become colnames, default to NA}

\item{colname}{the name of the column in the transposed data frame that will contain the old column names. Default to colnames}
}
\value{
The transposed dataframe
}
\description{
Pipe compatible transpose
}
