% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors}
\alias{colors}
\title{color Function}
\usage{
colors(print = FALSE)
}
\arguments{
\item{print}{If TRUE, print the palette names}
}
\value{
a vector with the palette names if print == TRUE
}
\description{
Print the palettes of RColorBrewer
}
\examples{
DailyHRB::colors()
}
