% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommonOrthologs.R
\name{getCommonOrthologs}
\alias{getCommonOrthologs}
\title{getCommonOrthologs Function}
\usage{
getCommonOrthologs(listMouse, listHuman, refMouse, refHuman)
}
\arguments{
\item{listMouse}{the genes from the mouse where we want the orthologs}

\item{listHuman}{the genes from the human where we want the orthologs}

\item{refMouse}{the gencode reference for mouse.
Either a number (the gencode version) or the location of the file.}

\item{refHuman}{the gencode reference for human.
Either a number (the gencode version) or the location of the file}
}
\description{
This function takes as input two count matrices from human and mouse and
return the matrices subset to the ortholog genes
}
\details{
When several genes from one species map to at least one gene from the other,
we return all genes. This function is inspired adapted from
}
