% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTheme.R
\name{my_theme}
\alias{my_theme}
\title{myTheme}
\usage{
my_theme(...)
}
\arguments{
\item{...}{Other arguments passed to the \link{theme} function in ggplot2.}
}
\value{
a theme
}
\description{
This function returns a theme that can be added to a ggplot object
It is based on the theme_classic(), with no white background so that
if the plot is saved as pdf, it incorporates better with slides. Also it avoids
doing theme_smth() + theme(other options)
}
